pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Mathieu Bellon <github@mathieubellon.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
supabase = "^0.7.1"
arrow = "^1.2.3"
pandas = "^1.5.1"
elasticsearch = "^8.5.0"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.src"
